+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerContainerFragment.kt
@@ -101,7 +101,7 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
                         theme.updateWindowStatusBar(it.window, StatusBarColor.Custom(ThemeColor.primaryUi01(Theme.ThemeType.DARK), true), it)
                     }
 
-                    upNextFragment.startTour()
+                    upNextFragment.onExpanded()
 
                     FirebaseAnalyticsTracker.openedUpNext()
                 } else if (newState == BottomSheetBehavior.STATE_COLLAPSED) {
@@ -109,6 +109,7 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
                     updateUpNextVisibility(false)
 
                     (activity as? FragmentHostListener)?.updateSystemColors()
+                    upNextFragment.onCollapsed()
                 }
             }
         })
@@ -210,9 +211,8 @@ class PlayerContainerFragment : BaseFragment(), HasBackstack {
     }
 
     fun openUpNext() {
-        binding?.let {
-            BottomSheetBehavior.from(it.upNextFrameBottomSheet).state = BottomSheetBehavior.STATE_EXPANDED
-        }
+        val upNextFragment = UpNextFragment.newInstance(source = UpNextSource.PLAYER)
+        (activity as? FragmentHostListener)?.showBottomSheet(upNextFragment)
     }
 
     fun openPlayer() {