+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/dao/EpisodeDao.kt
@@ -204,10 +204,10 @@ abstract class EpisodeDao {
     abstract fun updateSizeInBytes(sizeInBytes: Long, uuid: String)
 
     @Query("UPDATE podcast_episodes SET download_url = :url WHERE uuid = :uuid")
-    abstract fun updateDownloadUrl(url: String, uuid: String)
+    abstract suspend fun updateDownloadUrl(url: String, uuid: String)
 
     @Query("UPDATE podcast_episodes SET download_task_id = :taskId WHERE uuid = :uuid")
-    abstract fun updateDownloadTaskId(uuid: String, taskId: String?)
+    abstract suspend fun updateDownloadTaskId(uuid: String, taskId: String?)
 
     @Query("UPDATE podcast_episodes SET last_download_attempt_date = :lastDownloadAttemptDate WHERE uuid = :uuid")
     abstract fun updateLastDownloadAttemptDate(lastDownloadAttemptDate: Date, uuid: String)