+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+7.49
+-----
+
+*   Bug Fixes:
+    *   Improve multiselect handling in Up Next queue
+        ([#1398](https://github.com/Automattic/pocket-casts-android/pull/1390))
+
 7.48
 -----
 
@@ -7,13 +14,12 @@
 *   Bug Fixes:
     *   Avoid memory leak when opening Up Next queue
         ([#1397](https://github.com/Automattic/pocket-casts-android/pull/1397))
-    *   Improved multiselect handling in Up Next queue
-        ([#1395](https://github.com/Automattic/pocket-casts-android/pull/1392))
     *   Improved the downloading of episode show notes
         ([#1390](https://github.com/Automattic/pocket-casts-android/pull/1390))
     *   Fixed bug where theme selection UI was not functioning as intended
         ([#1399](https://github.com/Automattic/pocket-casts-android/pull/1399))
 
+        
 7.47
 -----
 
@@ -35,7 +41,7 @@
 *   Updates:
     *   Removed login with email and password on WearOS app
         ([#1356](https://github.com/Automattic/pocket-casts-android/pull/1356))
-    
+
 7.46.1
 ------
 *   Bug Fixes: